Reworked firing position
The punch now originates from a pin in the dwarf's arm. This makes it more consistent for calculating the direction iit's fired from. Likely need to make a specific direction vector specifically for the dwarf himself, and generate what happens to other targets using a circle to simulate the explosion.