Particle Systems Implemented
added two particle systems to the dwarf, and added a few lines of code to allow them to work. Implemented the PS's as prefabs that are dropped into the script to allow for modularity (in case we change the gauntlet or something), and allow a variable number of them on each game object (if added as a component to an object, you can only have one). Folders in Assets contain everything